llm-testing/requirements.txt
pydantic>=2.9.0
openai>=1.42.0
requests>=2.32.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
